Remove retired perl feature from UMDP3 checker (#42)
The UMDP3 checker uses a perl feature which assumes that keys() can be
called in a scalar context, based on an experimental feature
introduced in v5.14 which has now been withdrawn.

This changes the code to use a temporary hash to make the context of
the call explict.  This allows the keys() function to be used to
generate a list of unique extracted names.