Replace optparse with argparse in suite_report.py (#34)
* Repalce optparse with argparse in suite_report.py

Remove the deprecated optparse module and replace it with an similar
implementation using argparse.

* Fix linting problems in suite_report.py

Find and fix various problems reported by pylint.

* Make suite_report.py flake8 compliant

Fixes everything but a single case where the options are either an
E129 warning about having the same indentation as the next line or an
E127 warning about over-indentation - and there doesn't seem to be a
format that satisfies the checker!

* Add argument completion to suite_report.py

Enable command line argument completion for the two directory options
if/when argcomplete is available.

* Refactor with f-strings in place of formats

To keep the linter happy, most of the format() calls with f-strings.
This supposedly offers better performance and is more in keeping with
python 3.

* Amend initial comment about script location

Reviewer has pointed out that the comment about the UM respository is
no longer correct.  The text has been amended to reference the
SimSys_Scripts github repo instead.

* Clean up one of the formats in suite_report

Following suggestion from reviewer to remove one of the small number
of remaining formats in the code.

* Fix another format line

Apply fix suggested by reviewer

* Fix copy-paste error spotted in review

* Add reviewer suggestion and re-lint

Review suggested a way to remove an unnecessary format.  A quick
re-lint turned up a few outstanding issues which have been fixed.