Update branch pattern somewhat
This improves the pattern used to match the branch to make it closer
to the official git definition.  The pattern does not need to include
all the defined by git check-ref-format because the branch name must
already be compliant with these in order to be created in the first
place.