Initial version of a git_bdiff module (#100)
* Initial version of a git_bdiff module

A simple module that wraps around git diff and generates a complete
list of files changed on a branch since it diverged from the parent.
This is intended to replace the functionality of fcm_bdiff.

* Add the ability to specify a repository directory

Allow the constructor to specify the path to a repository and update
the run_git method to change to the directory before running each git
command.

* Include review changes from @ericaneininger

* Include review changes from @r-sharp

* Improve tests of error handling

Add string comparisons to existing exception tests and add a couple of
exception handling tests for the git_run method to improve testing
coverage.

* Improve hash pattern match

Reviewer has recommended changing the pattern match to require a full
40 character hex string.

* Minor change to remove some dead test code

* Update branch pattern somewhat

This improves the pattern used to match the branch to make it closer
to the official git definition.  The pattern does not need to include
all the defined by git check-ref-format because the branch name must
already be compliant with these in order to be created in the first
place.

---------

Co-authored-by: Erica Neininger <107684099+ericaneininger@users.noreply.github.com>